\name{SPR}

\alias{SPR}
\alias{SPR-package}

\docType{package}

\title{Shortest Path Algorithms with R}

\description{
SPR is a package of shortest path algorithms.
}

\details{
\tabular{ll}{
Package: \tab SPR\cr
Type: \tab Package\cr
Version: \tab 1.1.0\cr
Date: \tab 2018-01-14\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab Yes\cr
Depends: \tab Rcpp, RcppArmadillo \cr }
}
\section{How to cite this package}{
 \preformatted{@Manual{OHara2015,
 author       = {Keith O'Hara},
 title        = {{SPR}: Shortest Path with R.},
 year         = {2018},
 note         = {R package version 1.1.0.}}}
\emph{}
}

\section{License}{
SPR is licensed under the GNU General Public License (GPL) version 2, or (at your option) any later version.
}

\author{Keith O'Hara}